requirements.txt
panel==1.0.4
PyYAML==6.0
schema
